fix(spurctld): harden and re-home the job_submit hook after review
Move the embedded Lua backend out of spur-core into spurctld so the Lua
interpreter and its vendored C toolchain ship only with the controller,
not with spurd, the CLI, or the FFI shared object. spur-core keeps the
shared outcome/whitelist/apply types and the shell backend.

Also address several submit-hook robustness gaps:
- run check_submission_size before the hook so a doomed oversized spec
  isn't first serialized, forked to a shell, and parsed into a Lua VM
- run the Lua hook under block_in_place, matching the shell path
- cap the user-facing rejection reason (full stderr still logged)
- validate configured hooks at startup and reconfigure (absolute, secure,
  executable regular file; Lua additionally compiled) instead of failing
  on the first submission
- refuse a hook that isn't root/controller-owned or is group/world-writable
- reject a hook-set gres that conflicts with an explicit --gpus request at
  submit time rather than deferring to schedule time
- report an output-cap breach as a distinct error instead of a timeout

Document the submit-only enforcement boundary (scontrol update does not
re-run the hook), the shell-""-clears vs Lua-nil-noop asymmetry, and the
per-node HA hook-drift caveat in the examples.
