+++ crates/spur-core/Cargo.toml
@@ -24,7 +24,6 @@ rand = { workspace = true }
 subtle = { workspace = true }
 whoami = "2"
 nix = { workspace = true }
-mlua = { workspace = true }
 
 [dev-dependencies]
 tempfile = "3"
