+++ examples/spur.conf
@@ -113,6 +113,15 @@ memlock = "unlimited"
 # (whitelisted fields only). Shell decides via exit code + JSON on stdout; Lua
 # defines slurm_job_submit(job_desc, submit_uid) (Slurm job_submit/lua parity).
 # If both are set, shell runs first, then Lua. See examples/hooks/job_submit.{sh,lua}.
+#
+# Operational notes for the submit hooks:
+#   - The hook script must be an absolute path, owned by root or the controller
+#     user, and not group/world-writable (it runs as the controller, which is
+#     usually root). A bad path/permission fails at startup, not at first submit.
+#   - Enforcement is at submission only. `scontrol update` does NOT re-run the
+#     hook, so a policy like "cap walltime" is not re-applied to later edits.
+#   - In an HA controller set the hook file is per-node config; keep the script
+#     identical on every controller or policy will change on failover.
 [hooks]
 # job_submit = "/etc/spur/job_submit.sh"
 # job_submit_lua = "/etc/spur/job_submit.lua"
